server/requirements.txt
fastapi==0.88.0
sqlalchemy==1.4.46
databases==0.7.0
mysqlclient==2.1.1
aiomysql==0.1.1
uvicorn==0.20.0
pydantic==1.10.4
bcrypt==4.0.1
python-multipart==0.0.5
PyMySQL==1.0.2
PyJWT==2.6.0
